Always pass size_t varargs to p11_hash_xxx() functions
https://bugzilla.redhat.com/show_bug.cgi?id=985421

+++ common/tests/test-hash.c
@@ -87,7 +87,7 @@ test_sha1_long (CuTest *cu)
 	CuAssertTrue (cu, input != NULL);
 	memset (input, 'a', 1000000);
 
-	p11_hash_sha1 (checksum, input, 1000000, NULL);
+	p11_hash_sha1 (checksum, input, (size_t)1000000, NULL);
 	CuAssertTrue (cu, memcmp (expected, checksum, P11_HASH_SHA1_LEN) == 0);
 
 	free (input);
@@ -139,12 +139,12 @@ test_murmur2 (CuTest *cu)
 
 	assert (sizeof (one) == P11_HASH_MURMUR3_LEN);
 
-	p11_hash_murmur3 ((unsigned char *)&one, "one", 3, NULL);
-	p11_hash_murmur3 ((unsigned char *)&two, "two", 3, NULL);
-	p11_hash_murmur3 ((unsigned char *)&four, "four", 4, NULL);
-	p11_hash_murmur3 ((unsigned char *)&seven, "seven", 5, NULL);
-	p11_hash_murmur3 ((unsigned char *)&eleven, "eleven", 6, NULL);
-	p11_hash_murmur3 ((unsigned char *)&split, "ele", 3, "ven", 3, NULL);
+	p11_hash_murmur3 ((unsigned char *)&one, "one", (size_t)3, NULL);
+	p11_hash_murmur3 ((unsigned char *)&two, "two", (size_t)3, NULL);
+	p11_hash_murmur3 ((unsigned char *)&four, "four", (size_t)4, NULL);
+	p11_hash_murmur3 ((unsigned char *)&seven, "seven", (size_t)5, NULL);
+	p11_hash_murmur3 ((unsigned char *)&eleven, "eleven", (size_t)6, NULL);
+	p11_hash_murmur3 ((unsigned char *)&split, "ele", (size_t)3, "ven", (size_t)3, NULL);
 
 	CuAssertTrue (cu, one != two);
 	CuAssertTrue (cu, one != four);
